backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
firebase-admin>=6.5,<7.0